bunch o changes, good luck reviewer
