+++ src/__init__.py
@@ -1,0 +1,29 @@
+# Surveillance Demo Package
+# This package contains all the classes for the surveillance demo application
+
+from .stream_source import StreamSource
+from .app_config import AppConfig
+from .video_source import VideoSource
+from .model_manager import ModelManager
+from .ui_manager import UIManager
+from .effect_processor import EffectProcessor, ASCIIEffect, FaceOverlayEffect, FaceBlackoutEffect
+from .gesture_recognizer import GestureRecognizer
+from .thread_manager import ThreadManager
+from .surveillance_demo import SurveillanceDemo
+from .main import main
+
+__all__ = [
+    'StreamSource',
+    'AppConfig', 
+    'VideoSource',
+    'ModelManager',
+    'UIManager',
+    'EffectProcessor',
+    'ASCIIEffect',
+    'FaceOverlayEffect',
+    'FaceBlackoutEffect',
+    'GestureRecognizer',
+    'ThreadManager',
+    'SurveillanceDemo',
+    'main'
+]
