+++ src/stream_source.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class StreamSource(Enum):
+    RTSP = "rtsp"
+    WEBCAM = "webcam"
+    VIDEO = "video"
